backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
mlflow==2.8.1
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless==4.6.0.66
numpy==1.24.3
seaborn>=0.11.0
Pillow==10.1.0
ultralytics==8.0.220
opencv-python==4.8.1.78



